CERTIFICADOS/requirements.txt
Flask
google-cloud-storage
google-cloud-bigquery
reportlab
